requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# Utilities
requests>=2.31.0

# Optional (for dataset generation)
qrcode[pil]>=7.4.0
roboflow>=1.1.0
